requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.10
Pillow==10.1.0
python-docx==0.8.11
openpyxl==3.10.10
reportlab==4.0.9
PyPDF2==3.0.1
python-decouple==3.8
